# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Eye Diagnosis API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type PatientGender = (typeof PatientGender)[keyof typeof PatientGender];

export const PatientGender = {
  male: "male",
  female: "female",
  other: "other",
} as const;

export interface Patient {
  id: number;
  name: string;
  age: number;
  gender: PatientGender;
  createdAt: string;
}

export type CreatePatientInputGender =
  (typeof CreatePatientInputGender)[keyof typeof CreatePatientInputGender];

export const CreatePatientInputGender = {
  male: "male",
  female: "female",
  other: "other",
} as const;

export interface CreatePatientInput {
  name: string;
  age: number;
  gender: CreatePatientInputGender;
}

export type DiseaseResultSeverity =
  (typeof DiseaseResultSeverity)[keyof typeof DiseaseResultSeverity];

export const DiseaseResultSeverity = {
  none: "none",
  mild: "mild",
  moderate: "moderate",
  severe: "severe",
} as const;

export interface DiseaseResult {
  name: string;
  detected: boolean;
  confidence: number;
  severity: DiseaseResultSeverity;
  findings: string;
}

export type AnalysisEyeSide =
  (typeof AnalysisEyeSide)[keyof typeof AnalysisEyeSide];

export const AnalysisEyeSide = {
  left: "left",
  right: "right",
  both: "both",
} as const;

export type AnalysisOverallRisk =
  (typeof AnalysisOverallRisk)[keyof typeof AnalysisOverallRisk];

export const AnalysisOverallRisk = {
  normal: "normal",
  low: "low",
  moderate: "moderate",
  high: "high",
  critical: "critical",
} as const;

export interface Analysis {
  id: number;
  patientId: number;
  eyeSide: AnalysisEyeSide;
  imageName: string;
  imageUrl?: string;
  imageQualityScore: number;
  imageQualityOk: boolean;
  overallRisk: AnalysisOverallRisk;
  diseases: DiseaseResult[];
  summary: string;
  recommendations: string;
  analysisTime: number;
  createdAt: string;
}

export type CreateAnalysisBodyEyeSide =
  (typeof CreateAnalysisBodyEyeSide)[keyof typeof CreateAnalysisBodyEyeSide];

export const CreateAnalysisBodyEyeSide = {
  left: "left",
  right: "right",
  both: "both",
} as const;

export type CreateAnalysisBody = {
  patientId: number;
  imageBase64: string;
  imageName: string;
  eyeSide: CreateAnalysisBodyEyeSide;
};