# lib/api-zod/src/generated/types/analysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Eye Diagnosis API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisEyeSide } from "./analysisEyeSide";
import type { AnalysisOverallRisk } from "./analysisOverallRisk";
import type { DiseaseResult } from "./diseaseResult";

export interface Analysis {
  id: number;
  patientId: number;
  eyeSide: AnalysisEyeSide;
  imageName: string;
  imageUrl?: string;
  imageQualityScore: number;
  imageQualityOk: boolean;
  overallRisk: AnalysisOverallRisk;
  diseases: DiseaseResult[];
  summary: string;
  recommendations: string;
  analysisTime: number;
  createdAt: Date;
}