# lib/api-zod/src/generated/types/createPatientInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Eye Diagnosis API
 * OpenAPI spec version: 0.1.0
 */
import type { CreatePatientInputGender } from "./createPatientInputGender";

export interface CreatePatientInput {
  name: string;
  age: number;
  gender: CreatePatientInputGender;
}