# lib/api-zod/src/generated/types/diseaseResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Eye Diagnosis API
 * OpenAPI spec version: 0.1.0
 */
import type { DiseaseResultSeverity } from "./diseaseResultSeverity";

export interface DiseaseResult {
  name: string;
  detected: boolean;
  confidence: number;
  severity: DiseaseResultSeverity;
  findings: string;
}